requirements.txt
attrs
torchmetrics
imageio
matplotlib
matplotlib-inline
numpy
opencv-python
pandas
Pillow
pydicom
requests
scikit-image
scipy
tqdm
wandb
torchio
pylibjpeg 
pylibjpeg-libjpeg 

